pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "critical-weight-analysis"
version = "1.0.0"
description = "Critical Weight Analysis & Cybersecurity for Transformer Models - Lambda Labs GPU Optimized"
requires-python = ">=3.12"
dependencies = [
    # Core ML/AI libraries
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "huggingface-hub>=0.19.0",

    # Data & computation
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "matplotlib>=3.8.0",
    "scipy>=1.11.0",

    # Configuration & CLI
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",

    # Utilities
    "tqdm>=4.66.0",
    "psutil>=5.9.0",
    "sentencepiece>=0.1.99",
    "nvidia-ml-py>=12.0.0",

    # NOTE: PyTorch installed separately with CUDA support
    # Use: uv add torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]
quantization = [
    "bitsandbytes>=0.41.0",
]
lambda-optimized = [
    "flash-attn>=2.0.0",  # For Lambda Labs GPU optimization
    "triton>=2.1.0",      # GPU kernel optimization
]

[project.scripts]
cwa = "cwa.cli.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"